Remove "self type parameter" of IValidationTask
This allowed a more precise return type for IValidationTask.getTaskType,
but in practice, nothing has used this so far. Almost all existing uses
were of the form IValidationTask<?>, and in the few cases that had a
proper type argument, the code doesn't use the extra type information.

The disadvantage is that it requires an extra type argument in every
place where a non-concrete IValidationTask is used. The type parameter
also has to be explicitly passed through all intermediate classes, e. g.
DynamicFormulaTask.

This doesn't seem worth the effort and complexity at the moment, so I'm
removing the parameter for now. This doesn't reduce type safety in any
concrete case (and in fact eliminates one unchecked cast). We can always
re-add the type parameter in the future if we have code that would
clearly benefit from it.