Update to JavaFX 22 again and try to work around the NSMenuFX error
The problematic listener is added by the
MenuToolkit.autoAddWindowMenuItems, so removing that call hopefully also
prevents the error. The disadvantage is that the Window menu will no
longer contain a list of all open windows as is normally expected for
macOS applications. This shouldn't be a big issue though, because
ProB 2 UI usually doesn't have many windows open at once.