Update voparser to remove validation expression sequential operator
voparser 0.3.0 no longer supports the sequential operator. In the latest
formalization of VOs, sequential composition is done by passing a
validation task/expression as an argument to another task. This will be
implemented in ProB 2 UI in the future as a new property of the relevant
tasks.

The previous implementation of the sequential operator was very
incomplete and behaved basically identically to the conjunction
operator. In particular, the trace of the left expression wasn't passed
to the right expression, only the state space. This will be handled
properly by the new implementation.