Throw an exception of FindStateCommand returns an unexpected result
This matches what most other execute implementations do.

The comment about a possible null result seems to be outdated/incorrect.
FindStateCommand always sets a non-null result when it completes. The
only way for the result to be null is when the command execution throws
an exception or the command hasn't been executed at all. I assume this
is an old comment from when interrupting a command didn't automatically
throw an exception - in that case, the result could have also been null
because the command was interrupted by the user.