Remove LTL formula/pattern save/load buttons (as discussed)
The file format is incompatible with the ones used by probcli, ProB
Tcl/Tk, and the Rodin plugin, so it can only be used to transfer LTL
formulas/patterns between ProB 2 UI projects. No other verification
types in ProB 2 UI have a feature like this, and it's not clear why LTL
specifically would need this feature.