Make lots of classes final to silence most javac this-escape warnings
This isn't a great solution. Most of the this-escapes are actually fine,
e. g. passing this into the bean parameter of property constructors, and
it would be nicer if we could ignore those in a general way. Also, this
workaround obviously only works for classes that have no subclasses.

But these classes here are not meant to be extended anyway, so making
them final is a good thing regardless.