Fix PO tasks in machine not matching those in ProofObligationView
Previously, the ProofObligationItems in the machine validationTasks list
were not the same objects the ones shown in the ProofObligationView.
This led to the VO manager incorrectly showing all POs as "not checked"
in most cases.

This also separates the PO reading from ProofObligationView, which
ensures that they will be updated properly even if for whatever reason
that view is never loaded.