Update exception handling in VisBController.executeEvent
VisBPerformClickCommand cannot throw an EvaluationException or
ExecuteOperationException - these exception types are left over from the
previous code that used the regular APIs for executing operations by
predicate.

It seems that VisBPerformClickCommand currently doesn't throw errors at
all (it just prints them and then return an empty list of transitions),
but if it did report errors, they would be thrown as a ProBError.