Remove a few SimulatorStage listeners that shouldn't be needed
The current project listener isn't needed because the current machine
is guaranteed to change when the user opens a different project. The
window showing listener isn't needed because the listener is already
called once when SimulatorStage is created (and afterwards the listener
runs normally every time the current machine changes).