Remove default description for projects created from files
This doesn't add any information, because the machine file path is also
stored as part of the machine in the project.

It's also easy to accidentally leave this placeholder description in a
project, which can be confusing, especially if later other machines are
added or even the original machine is removed from the project.