Switch from javax.annotation to jakarta.annotation
We didn't actually have a proper dependency on javax.annotation - it was
pulled in indirectly via Guava, and apparently that indirect dependency
has disappeared in Guava 33.5.0, so this change is necessary for
updating Guava. It's also necessary for updating prob-java to Guice 7
(which no longer supports javax.* annotations) at some point.